Remove integrals2model.py
